DNN-10617 Code review and optimizations
