+++ src/Modules/Settings/Dnn.PersonaBar.Prompt/Prompt.Web/src/components/Output/util.js
@@ -1,3 +1,5 @@
+import React  from "react";
+
 export function formatLabel(input) {
     if (typeof input === "string") {
         // format camelcase and remove Is from labels
@@ -6,4 +8,32 @@ export function formatLabel(input) {
         return output;
     }
     return "";
+}
+
+export function getColumnsFromRow(row) {
+    const columns = [];
+    for (let key in row) {
+        if (!key.startsWith("__")) {
+            columns.push(key);
+        }
+    }
+    return columns;
+}
+
+export function renderObject(data, fieldOrder) {
+
+    const columns = !fieldOrder || fieldOrder.length == 0 ? getColumnsFromRow(data) : fieldOrder;
+    const rows = columns.map((fldName, index) => {
+        const lbl = formatLabel(fldName);
+        const fldVal = data[fldName] ? data[fldName] : '';
+        const cmd = data["__" + fldName] ? data["__" + fldName] : null;
+
+        if (cmd) {
+            return <tr key={index}><td className="dnn-prompt-lbl">{lbl}</td><td>:</td><td><a href="#" className="dnn-prompt-cmd-insert" data-cmd={cmd} title={cmd.replace(/'/g, '&quot;')}>{fldVal}</a></td></tr>;
+        } else {
+            return <tr key={index}><td className="dnn-prompt-lbl">{lbl}</td><td>:</td><td>{fldVal}</td></tr>;
+        }
+
+    });
+    return <table className="dnn-prompt-tbl"><tbody>{rows}</tbody></table>;
 }
